+++ src/CommandCenter/Controllers/LandingPageController.cs
@@ -11,6 +11,7 @@ namespace CommandCenter.Controllers
     using CommandCenter.Authorization;
     using CommandCenter.Marketplace;
     using CommandCenter.Models;
+    using Microsoft.AspNetCore.Authentication.OpenIdConnect;
     using Microsoft.AspNetCore.Authorization;
     using Microsoft.AspNetCore.Mvc;
     using Microsoft.Extensions.Logging;
@@ -21,7 +22,8 @@ namespace CommandCenter.Controllers
     /// <summary>
     /// Landing page.
     /// </summary>
-    [Authorize]
+    [Authorize(AuthenticationSchemes = OpenIdConnectDefaults.AuthenticationScheme)]
+    // Specify the auth scheme to be used for logging on users. This is for supporting WebAPI auth
     public class LandingPageController : Controller
     {
         private readonly ILogger<LandingPageController> logger;