+++ CHANGELOG.md
@@ -1,3 +1,10 @@
 # Change log
 
 1- [4-13-2020] Moving ths source from the initial location, https://github.com/Ercenk/ContosoAMPBasic to its new home
+
+2- [11-6-2020] Move to .NET 5.0
+
+3- [12-3-2020] 
+- Add Azure Queues as an option for handling notifications in addition to emails.
+- Implement token validation using the new Microsoft.Identity.Web library.
+- Added WebHookTokenParameters section to configuration to drive token validation